pyproject.toml
[tool.poetry]
name = "csc201ja22demo"
version = "0.1.0"
description = ""
authors = ["PHBrown <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
pylsp-mypy = "^0.5.7"
"graphics.py" = "^5.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
